Tidy assay data column names

Before performing differential expression genes analysis on expression data, it is important to ensure that the column names of the assay data is same with sample ID of the clinical or metadata. This function therefore help to transform the assay data column names to suit the sample name in the metadata.

transformMatrixCol(expMat, Sep, nSep, pos)

Arguments

expMat

A matrix data frame.

Sep

The delimiter that joins the appendage to desired column name

nSep

The count of the delimiter in the each column name.

pos

The position of the desired column name in Assay column name

Value

a data frame, with desired column name.

Details

Title

Examples

transformMatrixCol(Exprdata, Sep = ".", nSep = 2, pos = 2)
#> Error in is.data.frame(expMat): object 'Exprdata' not found